Refactor: MemberService 리팩토링 (#21)
