+++ src/main/java/org/sesac/slopedbe/member/controller/MemberController.java
@@ -1,6 +1,5 @@
 package org.sesac.slopedbe.member.controller;
 
-import org.sesac.slopedbe.auth.exception.MemberAlreadyExistsException;
 import org.sesac.slopedbe.auth.model.dto.MailVerificationRequest;
 import org.sesac.slopedbe.member.model.dto.UpdateRequest;
 import org.sesac.slopedbe.member.model.dto.request.CheckDuplicateIdRequest;
@@ -12,7 +11,6 @@ import org.sesac.slopedbe.member.service.MemberService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -39,28 +37,23 @@ public class MemberController {
     @PostMapping("/find-id")
     public ResponseEntity<String> findIdByEmail(@Valid @RequestBody MailVerificationRequest mailVerificationRequest) {
         String email = mailVerificationRequest.email();
-
-        try {
-            String id = memberService.findIdByEmail(email);
-            return ResponseEntity.ok(id);
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity.badRequest().body("Invalid email or verification code");
-        }
+        String id = memberService.findIdByEmail(email);
+        return ResponseEntity.ok(id);
     }
 
-    @PutMapping("/{id}")
+    @PutMapping("")
     public ResponseEntity<Member> updateMemberInfo(@RequestParam String email, @RequestParam String newNickname, @RequestParam String newPassword, boolean newDisability) {
         Member updatedMember = memberService.updateMemberInfo(email, newNickname, newPassword, newDisability);
         return ResponseEntity.ok(updatedMember);
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping("")
     public ResponseEntity<Void> deleteMember(@RequestParam String email) {
         memberService.deleteMember(email);
         return ResponseEntity.noContent().build();
     }
 
-    @PutMapping("/{id}/blacklist")
+    @PutMapping("/blacklist")
     public ResponseEntity<Member> updateStatus(@RequestParam String email, @RequestParam MemberStatus status) {
         Member updatedMember = memberService.updateMemberStatus(email, status);
         return ResponseEntity.ok(updatedMember);
@@ -74,20 +67,7 @@ public class MemberController {
 
     @PutMapping("/request-reset")
     public ResponseEntity<Member> updatePassword(@RequestBody UpdateRequest updateRequest){
-        String id = updateRequest.getId();
-        String newPassword = updateRequest.getNewPassword();
-
-        try {
-            Member updatedMember = memberService.updateMemberPassword(id, newPassword);
-            return ResponseEntity.ok(updatedMember);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null);
-        }
+        Member updatedMember = memberService.updateMemberPassword(updateRequest.getId(), updateRequest.getNewPassword());
+        return ResponseEntity.ok(updatedMember);
     }
-
-    @ExceptionHandler(MemberAlreadyExistsException.class)
-    public ResponseEntity<String> handleMemberAlreadyExistsException(MemberAlreadyExistsException ex) {
-        return new ResponseEntity<>(ex.getMessage(), HttpStatus.CONFLICT);
-    }
-
 }
