+++ src/main/java/org/sesac/slopedbe/member/model/dto/request/RegisterMemberRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record RegisterMemberRequest (
 	@NotBlank(message = "아이디를 입력해주세요")
-	String userId,
+	String id,
 	@NotBlank(message = "비밀번호를 입력해주세요")
 	String password,
 	@NotBlank(message = "이메일을 입력해주세요")
